Rename base representatives class to indicate it is a resource
